Create hui-hase-editor
