+++ src/panels/lovelace/strategies/types.ts
@@ -3,7 +3,7 @@ import { LovelaceStrategyConfig } from "../../../data/lovelace/config/strategy";
 import { LovelaceConfig } from "../../../data/lovelace/config/types";
 import { LovelaceViewConfig } from "../../../data/lovelace/config/view";
 import { HomeAssistant } from "../../../types";
-import { LovelaceGenericElementEditor } from "../types";
+import { LovelaceGenericEditor } from "../types";
 
 export type LovelaceStrategy<T = any> = {
   generate(config: LovelaceStrategyConfig, hass: HomeAssistant): Promise<T>;
@@ -21,6 +21,6 @@ export interface LovelaceViewStrategy
 export interface LovelaceSectionStrategy
   extends LovelaceStrategy<LovelaceSectionConfig> {}
 
-export interface LovelaceStrategyEditor extends LovelaceGenericElementEditor {
+export interface LovelaceStrategyEditor extends LovelaceGenericEditor {
   setConfig(config: LovelaceStrategyConfig): void;
 }
